# procede_api/src/swagger.json
{
    "swagger" : "2.0",
    "info" : {
        "version" : "1.0.0",
        "title" : "KillerBee",
        "description" : "API DOC",
        "license" : {
            "name" : "MIT",
            "url" : "https://opensource.org/licenses/MIT"
        }
    },
    "paths" : {
        "/createprocede" : {
            "post" : {
                "tags" : ["createProcede"],
                "summary" : "Create a new procede", 
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "body",
                        "schema" : {
                            "$ref" : "#definitions/Procede"
                        }
                    }
                ]
            }
        },
        "/getAllProcede" : {
            "get" : {
                "tags" : ["getAllProcede"],
                "summary" : "Get all the procede name", 
                "responses" : {
                    "200" : {
                        "description" : "OK",
                        "schema" : {
                            "$ref" : "#definitions/OK"
                        }
                    }
                }
            }
        },
        "/getprocede/:name" : {
            "get" : {
                "tags" : ["getProcede"],
                "summary" : "Get information about a procede",
                "responses" : {
                    "200" : {
                        "description" : "OK",
                        "schema" : {
                            "$ref" : "#definitions/Procede"
                        }
                    }
                }
            }
        },
        "/modify/:name" : {
            "put" : {
                "tags" : ["modifyProcede"],
                "summary" : "Modigy an existing procede",
                "parameters" : [
                    {
                        "in" : "body",
                        "name" : "body",
                        "schema" : {
                            "$ref" : "#definitions/Put"
                        }
                    }
                ]
            }
        },
        "/delete/:name" : {
            "delete" : {
                "tags" : ["deleteProcede"],
                "summary" : "Delete an existing procede"
            }
        }
    },

    "definitions" : {
        "Procede" : {
            "properties" : {
                "name": { "type": "string"},
                "description": { "type": "string" },
                "modele_freezbe": { "type": "string" },
                "tests": { "type": "string" }
            }
        },
        "OK" : {
            "properties" : {
                "name": { "type": "string"}
            }
        },
        "Put" : {
            "properties" : {
                "modelInfos" : {
                    "$ref" : "#definitions/Procede"
                }
            }
        }
    }
}